refactor: generateLottoNumbers를 domain으로 분리
