+++ README.md
@@ -12,8 +12,8 @@
    - [x] 1 ~ 45 사이의 숫자로 입력한다.
    - [x] 중복없이 숫자를 입력한다.
 4. [입력] 보너스 번호를 입력받는다.
-   - 1 ~ 45 사이의 숫자로 입력한다.
-   - 입력한 당첨 번호와 중복되면 안된다.
+   - [x] 1 ~ 45 사이의 숫자로 입력한다.
+   - [x] 입력한 당첨 번호와 중복되면 안된다.
 5. 사용자가 구매한 로또 번호와 당첨 번호를 비교하여 당첨 내역 및 수익률을 출력한다.
    - 1 ~ 5등까지 계산한다. (일치 여부 확인)
    - ```
